+++ src/scip/cons_indicator.c
@@ -7210,7 +7210,7 @@ SCIP_DECL_CONSGETPERMSYMGRAPH(consGetPermsymGraphIndicator)
    consnodeidx = SCIPaddSymgraphConsnode(scip, graph, cons, lhs, rhs);
 
    /* create nodes and edges for activation of constraint */
-   eqnodeidx = SCIPaddSymgraphOpnode(scip, graph, SYM_CONSOPTYPE_EQ);
+   eqnodeidx = SCIPaddSymgraphOpnode(scip, graph, (int) SYM_CONSOPTYPE_EQ); /*lint !e641*/
    SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, consnodeidx, eqnodeidx, FALSE, 0.0) );
 
    nodeidx = SCIPaddSymgraphValnode(scip, graph, (SCIP_Real) consdata->activeone);
@@ -7232,7 +7232,7 @@ SCIP_DECL_CONSGETPERMSYMGRAPH(consGetPermsymGraphIndicator)
    if( nlocvars > 1 || !SCIPisEQ(scip, vals[0], 1.0) || !SCIPisZero(scip, constant) )
    {
       /* encode aggregation by a sum-expression and connect it to bdexpr node */
-      opnodeidx = SCIPaddSymgraphOpnode(scip, graph, SYM_CONSOPTYPE_SUM);
+      opnodeidx = SCIPaddSymgraphOpnode(scip, graph, (int) SYM_CONSOPTYPE_SUM); /*lint !e641*/
       SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, eqnodeidx, opnodeidx, FALSE, 0.0) );
 
       /* add nodes and edges for variables in aggregation */
@@ -7244,7 +7244,7 @@ SCIP_DECL_CONSGETPERMSYMGRAPH(consGetPermsymGraphIndicator)
       SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, eqnodeidx, nodeidx, FALSE, 0.0) );
    }
 
-   slacknodeidx = SCIPaddSymgraphOpnode(scip, graph, SYM_CONSOPTYPE_SLACK);
+   slacknodeidx = SCIPaddSymgraphOpnode(scip, graph, (int) SYM_CONSOPTYPE_SLACK); /*lint !e641*/
    SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, consnodeidx, slacknodeidx, FALSE, 0.0) );
 
    /* create nodes and edges for (possibly aggregated) slack variable */
@@ -7258,7 +7258,7 @@ SCIP_DECL_CONSGETPERMSYMGRAPH(consGetPermsymGraphIndicator)
    if( nlocvars > 1 || !SCIPisEQ(scip, vals[0], 1.0) || !SCIPisZero(scip, constant) )
    {
       /* encode aggregation by a sum-expression and connect it to root node */
-      opnodeidx = SCIPaddSymgraphOpnode(scip, graph, SYM_CONSOPTYPE_SUM);
+      opnodeidx = SCIPaddSymgraphOpnode(scip, graph, (int) SYM_CONSOPTYPE_SUM); /*lint !e641*/
       SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, slacknodeidx, opnodeidx, FALSE, 0.0) );
 
       /* add nodes and edges for variables in aggregation */
@@ -7280,7 +7280,7 @@ SCIP_DECL_CONSGETPERMSYMGRAPH(consGetPermsymGraphIndicator)
    }
    nlocvars = nvarslincons;
 
-   opnodeidx = SCIPaddSymgraphOpnode(scip, graph, SYM_CONSOPTYPE_SUM);
+   opnodeidx = SCIPaddSymgraphOpnode(scip, graph, (int) SYM_CONSOPTYPE_SUM); /*lint !e641*/
    SCIP_CALL( SCIPaddSymgraphEdge(scip, graph, consnodeidx, opnodeidx, FALSE, 0.0) );
 
    SCIP_CALL( SCIPgetActiveVariables(scip, &vars, &vals, &nlocvars, &constant, SCIPisTransformed(scip)) );
